// src/lib/metadata-scraper/types.ts
/**
 * MIT License
 *
 * Copyright (c) 2021 Maximilian Schiller
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

export interface MetaData {
    title?: string;
    description?: string;
    icon?: string;
    image?: string;
    keywords?: string[];
    language?: string;
    type?: string;
    url?: string;
    provider?: string;

    [x: string]: string | string[] | undefined;
}

export type MetadataRule = [
    string,
    (el: Element) => (string | null),
]

export interface Context {
    url: string;
    options: Options;
}

export interface RuleSet {
    rules: MetadataRule[];
    defaultValue?: (context: Context) => string | string[];
    scorer?: (el: Element, score: any) => any;
    processor?: ((input: any, context: Context) => any);
}

export interface Options {
    maxRedirects?: number;
    ua?: string;
    lang?: string;
    timeout?: number;
    forceImageHttps?: boolean;
    html?: string;
    url?: string;
    customRules?: Record<string, RuleSet>;
}